Small admin services, app, registration refactoring & fixed message-bar double appear
