+++ src/app/shared/services/ministry-admin/ministry-admin.service.ts
@@ -5,7 +5,7 @@ import { Observable } from 'rxjs';
 import { AdminRoles } from 'shared/enum/admins';
 import { MinistryAdmin, MinistryAdminParameters } from 'shared/models/ministry-admin.model';
 import { SearchResponse } from 'shared/models/search.model';
-import { BaseAdminService } from '../base-admin/base-admin';
+import { BaseAdminService } from '../base-admin/base-admin.service';
 
 @Injectable({
   providedIn: 'root'
@@ -73,7 +73,7 @@ export class MinistryAdminService extends BaseAdminService {
    * This method reinvite Ministry Admin
    * @param adminId: string
    */
-  public reinviteAdmin(adminId: string): Observable<null> {
+  public reinviteAdmin(adminId: string): Observable<void> {
     return super.reinviteAdmin(adminId);
   }
 }
